pyproject.toml
[project]
name = "livekit-voice-agent"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "livekit-agents[openai,silero,turn-detector]~=1.2",
    "livekit-plugins-noise-cancellation~=0.2",
    "python-dotenv>=1.1.1",
    "bcrypt>=4.0.0",
    "Flask>=2.2",  # added Flask
]

[tool.uv.workspace]
members = [
    "livekit-voice-agent",
]
